+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* server/migrate/cheshire_json: Don't mangle chapter headings
 * client/text: Remember position in book last scrolled to
 * server/region/suspension: Redefine suspensions, ignoring sentence breaks in quotes before
 * server: Bump PyYAML for CVE-2017-18342